Separate into a client/common package
