+++ fabric-particles-v1/src/main/java/net/fabricmc/fabric/impl/client/particles/VanillaParticleManager.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.impl.particles;
+package net.fabricmc.fabric.impl.client.particles;
 
 import it.unimi.dsi.fastutil.ints.Int2ObjectMap;
 import net.minecraft.client.particle.ParticleFactory;
