+++ fabric-testmods/java/net/fabricmc/fabric/particles/ParticleModClient.java
@@ -20,9 +20,9 @@ import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.fabricmc.api.ModInitializer;
-import net.fabricmc.fabric.api.particles.ParticleFactoryRegistry;
+import net.fabricmc.fabric.api.client.particles.FabricSpriteProvider;
+import net.fabricmc.fabric.api.client.particles.ParticleFactoryRegistry;
 import net.fabricmc.fabric.api.particles.FabricParticleTypes;
-import net.fabricmc.fabric.api.particles.FabricSpriteProvider;
 import net.minecraft.block.Blocks;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.particle.AnimatedParticle;
